Move resize algorithm into preprocess dir
