+++ src/bindings/js/node/include/preprocess/resize_algorithm.hpp
